% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{devtoolverse_deps}
\alias{devtoolverse_deps}
\title{List all devtoolverse dependencies}
\usage{
devtoolverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
devtoolverse packages.}
}
\description{
List all devtoolverse dependencies
}
