% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{devtoolverse_packages}
\alias{devtoolverse_packages}
\title{List all packages in the devtoolverse}
\usage{
devtoolverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include devtoolverse in the list?}
}
\description{
List all packages in the devtoolverse
}
\examples{
devtoolverse_packages()
}
