% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{devtoolverse_update}
\alias{devtoolverse_update}
\title{Update devtoolverse packages}
\usage{
devtoolverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
devtoolverse packages.}
}
\description{
This will check to see if all devtoolverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
devtoolverse_update()
}
}
