% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{rpkgtools_deps}
\alias{rpkgtools_deps}
\title{List all rpkgtools dependencies}
\usage{
rpkgtools_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
rpkgtools packages.}
}
\description{
List all rpkgtools dependencies
}
