% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{rpkgtools_update}
\alias{rpkgtools_update}
\title{Update rpkgtools packages}
\usage{
rpkgtools_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
rpkgtools packages.}
}
\description{
This will check to see if all rpkgtools packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
rpkgtools_update()
}
}
